requirements.txt
arrow==0.12.0
b2==1.3.8
backports.functools-lru-cache==1.2.1
certifi==2018.11.29
chardet==3.0.4
duplicity==0.7.18.2
fasteners==0.14.1
funcsigs==1.0.2
futures==3.2.0
idna==2.8
logfury==0.1.2
monotonic==1.5
python-dateutil==2.8.0
PyYAML==3.13
requests==2.21.0
six==1.12.0
tqdm==4.31.1
urllib3==1.24.1
